requirements.txt
discord.py
requests
python-dotenv
json
pathlib